@@ -14,7 +14,7 @@
     <name>Spring-Blog</name>
     <description>Spring-Blog</description>
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
     </properties>
     <dependencies>
         <dependency>
