@@ -2,21 +2,29 @@
     <v-md-editor class="md-editor" v-model="essay.blogContent" :disabled-menus="[]" @upload-image="handleUploadImage"
         @save="UpdateEssay"></v-md-editor>
 </template>
-  
+
 <script>
-import { reqGetEssayById, reqUploadOssFile, reqUpdateEssay } from '@/api/blog'
+import {reqGetEssayById, reqUploadOssFile, reqUpdateEssay,} from '@/api/blog'
 export default {
     data() {
         return {
             essay: '',
             listLoading: true,
+            timer:'',
         }
     },
     created() {
-        this.GetEssayListByPage()
+      //监听路由变化 关闭定时器
+      this.$router.beforeEach( async (to, from, next) => {
+        this.clearTime()
+        next()
+      })
+        this.GetEssayListByID()
+        //启动定时保存定时器
+        this.startTime()
     },
     methods: {
-        async GetEssayListByPage() {
+        async GetEssayListByID() {
             this.listLoading = true
             const result = await reqGetEssayById(this.$route.query.blogId)
             if (result.code === 20000) {
@@ -48,19 +56,94 @@ export default {
             }
         },
         async UpdateEssay() {
+          //提示是否清空内容
+          if (!this.essay.blogContent.length){
+            this.$confirm('此操作将清空内容', '提示', {
+              confirmButtonText: '确定',
+              cancelButtonText: '取消',
+              type: 'warning'
+            }).then(async () => {
+              const result = await reqUpdateEssay(this.essay)
+              if (result.code === 20000) {
+                this.$message({
+                  type: 'success',
+                  message: result.msg
+                });
+                await this.GetEssayListByID()
+              } else {
+                this.$message({
+                  type: 'error',
+                  message: result.msg
+                });
+              }
+            }).catch(() => {
+              this.$message({
+                type: 'info',
+                message: '取消成功'
+              });
+              this.GetEssayListByID()
+            });
+          }else{
+            const inSpectBlog = await reqGetEssayById(this.$route.query.blogId)
+            //判断当前字符数量是否小于保存数量(500个字符)
+            if (this.essay.blogContent.length<inSpectBlog.data.blogContent.length-500){
+              this.$message({
+                type: 'info',
+                message: '为确保误删,你不能删除文章过多内容!'
+              });
+              await this.GetEssayListByID();
+              return;
+            }
             const result = await reqUpdateEssay(this.essay)
             if (result.code === 20000) {
-                this.$message({
-                    type: 'success',
-                    message: result.msg
-                });
-                this.GetEssayListByPage()
+              this.$message({
+                type: 'success',
+                message: result.msg
+              });
+              await this.GetEssayListByID()
             } else {
-                this.$message({
-                    type: 'error',
-                    message: result.msg
-                });
+              this.$message({
+                type: 'error',
+                message: result.msg
+              });
             }
+          }
+        },
+        async startTime(){
+          this.timer=setInterval(async ()=>{
+            this.$message({
+              type: 'success',
+              message: '定时保存中...'
+            });
+            const inSpectBlog = await reqGetEssayById(this.$route.query.blogId)
+            //定时保存字符必须大于之前字符数
+            console.log(this.essay.blogContent.length)
+            console.log(inSpectBlog.data.blogContent.length)
+            if (this.essay.blogContent.length<inSpectBlog.data.blogContent.length){
+              this.$message({
+                type: 'error',
+                message: '定时保存失败(请手动保存)!'
+              });
+              return;
+            }else{
+              const result = await reqUpdateEssay(this.essay)
+              if (result.code === 20000) {
+                this.$message({
+                  type: 'success',
+                  message: result.msg
+                });
+                await this.GetEssayListByID()
+              } else {
+                this.$message({
+                  type: 'error',
+                  message: result.msg
+                });
+              }
+            }
+          },5* 60 * 1000)
+        },
+      clearTime(){
+          clearInterval(this.timer)
         }
     },
 }
@@ -158,4 +241,3 @@ export default {
     }
 }
 </style>
-  
