@@ -7,6 +7,7 @@ import com.hncpu.blog.entity.TbBlogEntity;
 import com.hncpu.blog.entity.TbClassifyEntity;
 import com.hncpu.blog.mapper.TbClassifyMapper;
 import com.hncpu.blog.service.TbBlogService;
+import com.hncpu.blog.service.TbClassIfyCountService;
 import com.hncpu.blog.service.TbClassifyService;
 import com.hncpu.blog.exception.BlogException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,7 +20,7 @@ import java.util.List;
 @Service
 public class TbClassifyServiceImpl extends ServiceImpl<TbClassifyMapper, TbClassifyEntity> implements TbClassifyService {
     @Autowired
-    TbBlogService tbBlogService;
+    private TbBlogService tbBlogService;
     /** 查询全部归档 */
     @Override
     public List<TbClassifyEntity> queryAll() {
